move todo al tree preloader
