+++ app/services/yml_loader.rb
@@ -18,7 +18,7 @@ class YmlLoader
     load_subjects
     load_prerequisites
     load_current_optional_subjects
-    PreloadedSubjects.reload!
+    TreePreloader.refresh_cache!
   end
 
   private
